django_test/requirements.txt
Django==1.11.3
psycopg2==2.7.1
djangorestframework
